refactor(6880): split task operations snapshot helpers
